[chore] 폴더 변경 및 의존성 변경 - #15
