+++ chat-service/src/main/java/org/kiru/chat/adapter/in/web/ChatRoomController.java
@@ -7,8 +7,8 @@ import org.kiru.chat.application.port.in.CreateRoomUseCase;
 import org.kiru.chat.application.port.in.GetChatRoomUseCase;
 import org.kiru.chat.application.port.in.SendMessageUseCase;
 import org.kiru.chat.config.argumentresolve.UserId;
-import org.kiru.core.chatroom.domain.ChatRoom;
-import org.kiru.core.message.domain.Message;
+import org.kiru.core.chat.chatroom.domain.ChatRoom;
+import org.kiru.core.chat.message.domain.Message;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
