+++ chat-service/src/main/java/org/kiru/chat/application/port/in/GetChatRoomUseCase.java
@@ -1,7 +1,7 @@
 package org.kiru.chat.application.port.in;
 
 import java.util.List;
-import org.kiru.core.chatroom.domain.ChatRoom;
+import org.kiru.core.chat.chatroom.domain.ChatRoom;
 
 public interface GetChatRoomUseCase {
     ChatRoom findRoomById(Long roomId, Long userId);
