+++ chat-service/src/main/java/org/kiru/chat/application/port/in/SendMessageUseCase.java
@@ -1,6 +1,6 @@
 package org.kiru.chat.application.port.in;
 
-import org.kiru.core.message.domain.Message;
+import org.kiru.core.chat.message.domain.Message;
 
 public interface SendMessageUseCase {
     Message sendMessage(Long roomId, Message message);
