+++ core/src/main/java/org/kiru/core/chat/chatroom/domain/ChatRoomType.java
@@ -1,0 +1,6 @@
+package org.kiru.core.chat.chatroom.domain;
+
+public enum ChatRoomType {
+    PRIVATE, // 1대1 채팅
+    GROUP    // 다중 채팅
+}
