--- core/src/main/java/org/kiru/core/chatroom/domain/ChatRoomType.java
@@ -1,6 +1,0 @@
-package org.kiru.core.chatroom.domain;
-
-public enum ChatRoomType {
-    PRIVATE, // 1대1 채팅
-    GROUP    // 다중 채팅
-}
