--- core/src/main/java/org/kiru/core/talent/entity/UserTalent.java
@@ -1,34 +1,0 @@
-package org.kiru.core.talent.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.kiru.core.talent.domain.Talent.*;
-
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-@Table(name = "user_talents")
-public class UserTalent {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotNull
-    private Long userId;
-
-    @Enumerated(EnumType.STRING)
-    @NotNull
-    private TalentType talentType;
-}
