--- core/src/main/java/org/kiru/core/user/domain/LoginType.java
@@ -1,5 +1,0 @@
-package org.kiru.core.user.domain;
-
-public enum LoginType {
-    KAKAO,APPLE,GOOGLE,INSTAGRAM
-}
