+++ core/src/main/java/org/kiru/core/user/user/domain/LoginType.java
@@ -1,0 +1,5 @@
+package org.kiru.core.user.user.domain;
+
+public enum LoginType {
+    KAKAO,APPLE,GOOGLE,INSTAGRAM,LOCAL
+}
