+++ core/src/main/java/org/kiru/core/user/userPurpose/domain/PurposeType.java
@@ -1,0 +1,9 @@
+package org.kiru.core.user.userPurpose.domain;
+
+public enum PurposeType {
+    GET_ALONG_WITH_U,
+    WANT_TO_COLLABORATE,
+    WANNA_MAKE_NEW_BRAND,
+    ART_RESIDENCY,
+    GROUP_EXHIBITION
+}
