+++ core/src/main/java/org/kiru/core/user/userlike/domain/LikeStatus.java
@@ -1,0 +1,5 @@
+package org.kiru.core.user.userlike.domain;
+
+public enum LikeStatus {
+    LIKE,DISLIKE
+}
