--- core/src/main/java/org/kiru/core/userPurpose/domain/PurposeType.java
@@ -1,9 +1,0 @@
-package org.kiru.core.userPurpose.domain;
-
-public enum PurposeType {
-    GET_ALONG_WITH_U,
-    WANT_TO_COLLABORATE,
-    WANNA_MAKE_NEW_BRAND,
-    ART_RESIDENCY,
-    GROUP_EXHIBITION
-}
