--- core/src/main/java/org/kiru/core/userlike/domain/LikeStatus.java
@@ -1,5 +1,0 @@
-package org.kiru.core.userlike.domain;
-
-public enum LikeStatus {
-    LIKE,DISLIKE
-}
