+++ settings.gradle
@@ -4,3 +4,4 @@ include ':core'
 include ':chat-service'
 include ':gateway-service'
 include ':user-service'
+include ':discovery'
